+++ GameFrameX.Hotfix/Logic/Server/Server/ServerComponentAgent.cs
@@ -34,7 +34,7 @@ namespace GameFrameX.Hotfix.Logic.Server.Server
             {
                 LogHelper.Debug($"ServerCompAgent.CrossDayTimeHandler.跨天定时器执行{TimeHelper.CurrentTimeWithFullString()}");
                 await ActorManager.RoleCrossDay(1);
-                await ActorManager.CrossDay(1, ActorType.Server);
+                await ActorManager.CrossDay(1, (ushort)ActorType.Server);
             }
         }
 
@@ -59,7 +59,7 @@ namespace GameFrameX.Hotfix.Logic.Server.Server
             OwnerComponent.OnlineSet.Remove(roleId);
             return ValueTask.CompletedTask;
         }
-        
+
         public static async Task OnlineRoleForeach(Action<PlayerComponentAgent> func)
         {
             var serverComp = await ActorManager.GetComponentAgent<ServerComponentAgent>();
